+++ src/main/java/com/justifiedsolutions/justpdf/pdf/contents/DeviceRGBOperator.java
@@ -3,13 +3,11 @@ package com.justifiedsolutions.justpdf.pdf.contents;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
-import java.util.Objects;
 
 /**
  * Superclass for operators that work on {@link DeviceRGB} colorspace.
  */
-abstract class DeviceRGBOperator implements ColorGraphicsOperator {
-    private final DeviceRGB colorSpace;
+abstract class DeviceRGBOperator extends DeviceColorSpaceOperator {
 
     /**
      * Creates a new operator that sets the stroke color in the RGB color space in a content stream.
@@ -17,33 +15,16 @@ abstract class DeviceRGBOperator implements ColorGraphicsOperator {
      * @param colorSpace the color space
      */
     protected DeviceRGBOperator(DeviceRGB colorSpace) {
-        this.colorSpace = Objects.requireNonNull(colorSpace);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(colorSpace);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        DeviceRGBOperator that = (DeviceRGBOperator) o;
-        return colorSpace.equals(that.colorSpace);
+        super(colorSpace);
     }
 
     @Override
     public void writeToPDF(OutputStream pdf) throws IOException {
-        colorSpace.getRed().writeToPDF(pdf);
+        getColorSpace().getRed().writeToPDF(pdf);
         pdf.write(' ');
-        colorSpace.getGreen().writeToPDF(pdf);
+        getColorSpace().getGreen().writeToPDF(pdf);
         pdf.write(' ');
-        colorSpace.getBlue().writeToPDF(pdf);
+        getColorSpace().getBlue().writeToPDF(pdf);
         pdf.write(' ');
         pdf.write(getOperatorCode().getBytes(StandardCharsets.US_ASCII));
         pdf.write('\n');
@@ -55,14 +36,6 @@ abstract class DeviceRGBOperator implements ColorGraphicsOperator {
      * @return the colorspace
      */
     protected DeviceRGB getColorSpace() {
-        return colorSpace;
+        return (DeviceRGB) getDeviceColorSpace();
     }
-
-    /**
-     * Gets the operator code for writing to a PDF.
-     *
-     * @return the code
-     */
-    protected abstract String getOperatorCode();
-
 }
